extract to styled files @ormiz
