+++ examples/client/Locomotion/src/Components/BusinessAccountText/index.tsx
@@ -1,25 +1,6 @@
 import React from 'react';
-import { Text, View } from 'react-native';
-import styled from 'styled-components';
+import { BoldTitle, SubTitle, TitleWithSubTitle } from './styled';
 
-const TitleWithSubTitle = styled(View)`
-  display: flex;
-  flex-direction: column;
-  flex: 1 0 0;
-`;
-const BaseText = styled(Text)`
-color: #212229;
-font-family: Montserrat;
-font-size: 14px;
-font-style: normal;
-line-height: 20px;
-`;
-const SubTitle = styled(BaseText)`
-font-weight: 500;
-`;
-const BoldTitle = styled(BaseText)`
-font-weight: 700;
-`;
 interface BusinessAccountTextProps {
     title: string;
     subTitle: string;
